Rename file with checkHexa
