+++ src/lib/components/FrontPage.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import { checkHexa } from '$lib/ts/checkHexa';
+	import { checkHexa } from '$lib/logic';
 
 	let inputfield: string = '';
